feat: 마이페이지 프로필 정보 조회

+++ src/main/java/hyundai/movie_review/tier/dto/TierMyPageInfoDto.java
@@ -12,12 +12,12 @@ public record TierMyPageInfoDto (
         @Schema(description = "티어 이미지 경로", example = "http://k.kakaocdn.net/...")
         String tierImage,
         @Schema(description = "현재 티어에서 다음 티어까지 필요한 총 점수", example = "100")
-        Integer tierRequiredPoints,
+        Long tierRequiredPoints,
         @Schema(description = "사용자의 점수", example = "30")
-        Integer tierCurrentPoints
+        Long tierCurrentPoints
 ){
-    public static TierMyPageInfoDto of(Tier tier, Integer score){
-        int requiredPoints = 0, currentPoints = 0;
+    public static TierMyPageInfoDto of(Tier tier, Long score){
+        long requiredPoints = 0, currentPoints = 0;
 
         // Lv 1 ~ 3은 다음 티어/레벨까지 100 포인트 필요
         if(tier.getId() == 1){ requiredPoints = 100; currentPoints = score; }
